Tek3/Web/Area/server/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pymysql==1.0.2
requests
pyjwt
python-dotenv
discord.py
asyncio